azure-rm-nginx-configuration - fix package_data
When a user uploads an nginx configuration as a tarball using `package_data`, the NGINXaaS API extracts and returns
the files. Terraform sees this on subsequent plans and thinks it needs to nullify these config files.

As a fix, this adds conditions to the Read and Update methods, like protected files, to ignore config_files
when using package_data.